# server/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Message": {
				"properties": {
					"id": {
						"type": "string"
					},
					"text": {
						"type": "string"
					},
					"category": {
						"type": "string",
						"enum": [
							"system",
							"user",
							"billing"
						]
					},
					"priority": {
						"type": "number",
						"format": "double"
					},
					"isRead": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"text",
					"category",
					"priority",
					"isRead",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorMessage": {
				"properties": {
					"error": {
						"type": "string"
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageCreateRequest": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"description": "Message API server utilizing Express, SQLite, tsoa, and Drizzle ORM.",
		"contact": {}
	},
	"paths": {
		"/api/v1/messages": {
			"get": {
				"operationId": "GetMessages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Message"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Fetch messages with support for search, categorization, range filtering, and custom sorting",
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"system",
								"user",
								"billing"
							]
						}
					},
					{
						"in": "query",
						"name": "minPriority",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "isRead",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"createdAt",
								"priority"
							]
						}
					},
					{
						"in": "query",
						"name": "order",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Message"
								}
							}
						}
					}
				},
				"description": "Create a new message",
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MessageCreateRequest"
							}
						}
					}
				}
			}
		},
		"/api/v1/messages/{id}": {
			"get": {
				"operationId": "GetMessageById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Message"
										},
										{
											"$ref": "#/components/schemas/ErrorMessage"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Message not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					}
				},
				"description": "Fetch a single message by ID",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Message"
										},
										{
											"$ref": "#/components/schemas/ErrorMessage"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Message not found to update",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					}
				},
				"description": "Update an existing message text description",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MessageCreateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/ErrorMessage"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Message not found to delete",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorMessage"
								}
							}
						}
					}
				},
				"description": "Remove a message",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}